requirements.txt
Django>=4.2
Pillow>=10.0
python-decouple>=3.8
dj-database-url>=2.1.0
django-crispy-forms>=2.0
crispy-tailwind>=0.5
whitenoise>=6.5
gunicorn>=21.0
